fix(admin): make event duplication reachable — extract events/[id]/duplicate.js (Closes #432)
POST /api/admin/events/{id}/duplicate 404'd in production: the duplicate
logic lived in events/[id].js behind `if (url.pathname.endsWith("/duplicate"))`,
but Cloudflare Pages route handlers match the path exactly, so a single-segment
[id].js never receives the /duplicate sub-path. The frontend (eventsApi.duplicate)
called it anyway; the unit test passed only by importing the handler directly,
bypassing routing.

- Add functions/api/admin/events/[id]/duplicate.js exporting onRequestPost,
  keyed on params.id via validateId (mirrors publish.js / archive.js / reveal-mode.js).
- Remove the now-dead onRequestPost (and its endsWith("/duplicate") branch) from
  events/[id].js. getEventId stays — PATCH/PUT/DELETE still use it.
- Repoint the two duplicate unit tests to the new handler (passing params.id).
- Re-document the path + EventDuplicateRequest/EventDuplicateResponse in
  api-spec.yaml and API_DOCUMENTATION.md. The #416 drift guard now enforces
  this: the new route file forces the spec entry (70 route files, 69 documented).

Verified in a clean node:22 Linux container: full backend suite 509 passed,
validate:openapi reports no drift.
