+++ docs/API_DOCUMENTATION.md
@@ -168,6 +168,7 @@ All admin routes require a valid session cookie. State-changing routes also requ
 | `/api/admin/events/{id}/publish`     | `POST`   | editor       | Body: `{ publish: boolean }`. Refuses to publish an event with no performances.                                                                                        |
 | `/api/admin/events/{id}/archive`     | `POST`   | admin        | Archives an event and clears `is_published`.                                                                                                                           |
 | `/api/admin/events/{id}/metrics`     | `GET`    | viewer       | Schedule-build analytics for one event.                                                                                                                                |
+| `/api/admin/events/{id}/duplicate`   | `POST`   | editor       | Body: `{ name, date, slug }`. Copies the event's performances into a new unpublished draft.                                                                            |
 | `/api/admin/events/wizard`           | `POST`   | admin        | One-shot creation of a draft event, its venues (find-or-create by name), and lineup. Body: `{ event, venues[], bands[] }`; each `bands[].venueIndex` indexes `venues`. |
 | `/api/admin/events/{id}/reveal-mode` | `POST`   | editor       | Body: `{ reveal_mode: boolean }`. Toggles the lineup embargo flag for the event.                                                                                       |
 | `/api/admin/venues`                  | `GET`    | viewer       | Lists venues with band counts. Contact details are redacted for viewers.                                                                                               |
