+++ functions/api/admin/events/[id].js
@@ -560,220 +560,6 @@ export async function onRequestPut(context) {
   }
 }
 
-// POST - Duplicate event (editor and admin only)
-export async function onRequestPost(context) {
-  const { request, env } = context;
-  const { DB } = env;
-  const url = new URL(request.url);
-  const ipAddress = getClientIP(request);
-
-  try {
-    // Check permission (editor and above)
-    const permCheck = await checkPermission(context, 'editor');
-    if (permCheck.error) {
-      return permCheck.response;
-    }
-
-    const currentUser = permCheck.user;
-    const eventId = getEventId(request);
-
-    if (!eventId || isNaN(eventId)) {
-      return new Response(
-        JSON.stringify({
-          error: "Bad request",
-          message: "Invalid event ID",
-        }),
-        {
-          status: 400,
-          headers: { "Content-Type": "application/json" },
-        },
-      );
-    }
-
-    // Check if this is a duplicate request
-    if (url.pathname.endsWith("/duplicate")) {
-      const body = await request.json().catch(() => ({}));
-      const { name, date, slug } = body;
-
-      // Validation
-      if (!name || !date || !slug) {
-        return new Response(
-          JSON.stringify({
-            error: "Validation error",
-            message: "Name, date, and slug are required for duplicate event",
-          }),
-          {
-            status: 400,
-            headers: { "Content-Type": "application/json" },
-          },
-        );
-      }
-
-      // Get original event
-      const originalEvent = await DB.prepare(
-        `
-        SELECT * FROM events WHERE id = ?
-      `,
-      )
-        .bind(eventId)
-        .first();
-
-      if (!originalEvent) {
-        return new Response(
-          JSON.stringify({
-            error: "Not found",
-            message: "Original event not found",
-          }),
-          {
-            status: 404,
-            headers: { "Content-Type": "application/json" },
-          },
-        );
-      }
-
-      // Check if slug already exists
-      const existingEvent = await DB.prepare(
-        `
-        SELECT id FROM events WHERE slug = ?
-      `,
-      )
-        .bind(slug)
-        .first();
-
-      if (existingEvent) {
-        return new Response(
-          JSON.stringify({
-            error: "Conflict",
-            message: "An event with this slug already exists",
-          }),
-          {
-            status: 409,
-            headers: { "Content-Type": "application/json" },
-          },
-        );
-      }
-
-      // Create new event (unpublished by default)
-      const newEvent = await DB.prepare(
-        `
-        INSERT INTO events (
-          name,
-          date,
-          slug,
-          status,
-          is_published,
-          description,
-          city,
-          ticket_url,
-          venue_info,
-          social_links,
-          theme_colors,
-          created_by_user_id
-        )
-        VALUES (?, ?, ?, ?, 0, ?, ?, ?, ?, ?, ?, ?)
-        RETURNING *
-      `,
-      )
-        .bind(
-          name,
-          date,
-          slug,
-          "draft",
-          originalEvent.description || null,
-          originalEvent.city || null,
-          originalEvent.ticket_url || null,
-          originalEvent.venue_info || null,
-          originalEvent.social_links || null,
-          originalEvent.theme_colors || null,
-          currentUser.userId,
-        )
-        .first();
-
-      if (!newEvent) {
-        throw new Error("events INSERT returned null");
-      }
-
-      // Copy all bands from original event (compensating delete if this fails)
-      try {
-        await DB.prepare(
-          `
-          INSERT INTO performances (event_id, venue_id, band_profile_id, start_time, end_time, notes)
-          SELECT ?, venue_id, band_profile_id, start_time, end_time, notes
-          FROM performances
-          WHERE event_id = ?
-        `,
-        )
-          .bind(newEvent.id, eventId)
-          .run();
-      } catch (copyError) {
-        await DB.prepare("DELETE FROM events WHERE id = ?").bind(newEvent.id).run();
-        throw copyError;
-      }
-
-      // Get band count
-      const bandCount = await DB.prepare(
-        `
-        SELECT COUNT(*) as count FROM performances WHERE event_id = ?
-      `,
-      )
-        .bind(newEvent.id)
-        .first();
-
-      // Audit log
-      await auditLog(
-        env,
-        currentUser.userId,
-        'event.duplicated',
-        'event',
-        newEvent.id,
-        {
-          name: newEvent.name,
-          originalEventId: eventId,
-          bandsCopied: bandCount.count,
-        },
-        ipAddress
-      );
-
-      return new Response(
-        JSON.stringify({
-          success: true,
-          event: newEvent,
-          bandsCopied: bandCount.count,
-          message: `Event duplicated successfully with ${bandCount.count} bands`,
-        }),
-        {
-          status: 201,
-          headers: { "Content-Type": "application/json" },
-        },
-      );
-    }
-
-    return new Response(
-      JSON.stringify({
-        error: "Not found",
-        message: "Unknown operation",
-      }),
-      {
-        status: 404,
-        headers: { "Content-Type": "application/json" },
-      },
-    );
-  } catch (error) {
-    console.error("Error duplicating event:", error);
-
-    return new Response(
-      JSON.stringify({
-        error: "Database error",
-        message: "Failed to duplicate event",
-      }),
-      {
-        status: 500,
-        headers: { "Content-Type": "application/json" },
-      },
-    );
-  }
-}
-
 // DELETE - Delete event (admin only)
 export async function onRequestDelete(context) {
   const { request, env } = context;
